+++ resources/views/schema/lottie.blade.php
@@ -1,15 +1,22 @@
 @php
-    /** @var \Visualbuilder\Lottie\Components\Lottie $component */
-    $component = $getComponent();
+    $src = $getSrc();
+    $autoplay = $getAutoplay();
+    $loop = $getLoop();
+    $speed = $getSpeed();
+    $width = $getWidth();
+    $height = $getHeight();
+    $trigger = $getTrigger();
+    $onComplete = $getOnComplete();
+    $respectReducedMotion = $getRespectReducedMotion();
 @endphp
 
 <dotlottie-wc
-    src="{{ $component->getSrc() }}"
-    @if ($component->getAutoplay()) autoplay @endif
-    @if ($component->getLoop()) loop @endif
-    @if ($component->getSpeed() !== null) speed="{{ $component->getSpeed() }}" @endif
-    style="width: {{ $component->getWidth() }}; height: {{ $component->getHeight() }};"
-    data-vb-lottie-trigger="{{ $component->getTrigger() }}"
-    @if ($component->getOnComplete()) data-vb-lottie-on-complete="{{ $component->getOnComplete() }}" @endif
-    @if ($component->getRespectReducedMotion()) data-vb-lottie-respect-reduced-motion="true" @endif
+    src="{{ $src }}"
+    @if ($autoplay) autoplay @endif
+    @if ($loop) loop @endif
+    @if ($speed !== null) speed="{{ $speed }}" @endif
+    style="width: {{ $width }}; height: {{ $height }};"
+    data-vb-lottie-trigger="{{ $trigger }}"
+    @if ($onComplete) data-vb-lottie-on-complete="{{ $onComplete }}" @endif
+    @if ($respectReducedMotion) data-vb-lottie-respect-reduced-motion="true" @endif
 ></dotlottie-wc>